Add validations, tweak styles
The hash will now be checked before sending a request to the API.
It must exist and pass a regex test for a request to be sent. Even
after that, if a hash is already present in the list, the request
will not be sent and a message will be displayed instead.